gameScreen object created
