+++ js/data/game-data.js
@@ -1,6 +1,4 @@
-import game1 from "../screens/game-1";
-import game2 from "../screens/game-2";
-import game3 from "../screens/game-3";
+import gameScreens from "./game-screens";
 
 /*
 1. Описание игры
@@ -26,145 +24,6 @@ const initialState = {
   lives: 3,
   time: 30
 };
-
-const gameScreens = [{
-  type: `tinder-like`,
-  question: `Угадай, фото или рисунок?`,
-  answers: [{
-    image: {
-      url: `https://k42.kn3.net/CF42609C8.jpg`
-    },
-    type: `painting`
-  }]
-}, {
-  type: `two-of-two`,
-  question: `Угадайте для каждого изображения фото или рисунок?`,
-  answers: [{
-      image: {
-        url: `https://k42.kn3.net/CF42609C8.jpg`
-      },
-      type: `photo`
-    },
-    {
-      image: {
-        url: `https://k42.kn3.net/CF42609C8.jpg`
-      },
-      type: `photo`
-    }
-  ]
-}, {
-  type: `tinder-like`,
-  question: `Угадай, фото или рисунок?`,
-  answers: [{
-    image: {
-      url: `https://k42.kn3.net/CF42609C8.jpg`
-    },
-    type: `photo`
-  }]
-}, {
-  type: `one-of-three`,
-  question: `Найдите рисунок среди изображений`,
-  answers: [{
-      image: {
-        url: `https://k42.kn3.net/CF42609C8.jpg`
-      },
-      type: `painting`
-    },
-    {
-      image: {
-        url: `https://k42.kn3.net/CF42609C8.jpg`
-      },
-      type: `painting`
-    },
-    {
-      image: {
-        url: `https://k42.kn3.net/CF42609C8.jpg`
-      },
-      type: `painting`
-    }
-  ]
-}, {
-  type: `tinder-like`,
-  question: `Угадай, фото или рисунок?`,
-  answers: [{
-    image: {
-      url: `https://k42.kn3.net/CF42609C8.jpg`
-    },
-    type: `painting`
-  }]
-}, {
-  type: `one-of-three`,
-  question: `Найдите рисунок среди изображений`,
-  answers: [{
-      image: {
-        url: `https://k42.kn3.net/CF42609C8.jpg`
-      },
-      type: `painting`
-    },
-    {
-      image: {
-        url: `https://k42.kn3.net/CF42609C8.jpg`
-      },
-      type: `painting`
-    },
-    {
-      image: {
-        url: `https://k42.kn3.net/CF42609C8.jpg`
-      },
-      type: `painting`
-    }
-  ]
-}, {
-  type: `two-of-two`,
-  question: `Угадайте для каждого изображения фото или рисунок?`,
-  answers: [{
-      image: {
-        url: `https://k42.kn3.net/CF42609C8.jpg`
-      },
-      type: `photo`
-    },
-    {
-      image: {
-        url: `https://k42.kn3.net/CF42609C8.jpg`
-      },
-      type: `photo`
-    }
-  ]
-}, {
-  type: `tinder-like`,
-  question: `Угадай, фото или рисунок?`,
-  answers: [{
-    image: {
-      url: `https://k42.kn3.net/CF42609C8.jpg`
-    },
-    type: `painting`
-  }]
-}, {
-  type: `two-of-two`,
-  question: `Угадайте для каждого изображения фото или рисунок?`,
-  answers: [{
-      image: {
-        url: `https://k42.kn3.net/CF42609C8.jpg`
-      },
-      type: `photo`
-    },
-    {
-      image: {
-        url: `https://k42.kn3.net/CF42609C8.jpg`
-      },
-      type: `photo`
-    }
-  ]
-}, {
-  type: `tinder-like`,
-  question: `Угадай, фото или рисунок?`,
-  answers: [{
-    image: {
-      url: `https://k42.kn3.net/CF42609C8.jpg`
-    },
-    type: `painting`
-  }]
-}];
 
 // const gamePlay = {
 //   GAME_SETUP: {
