+++ js/data/game-screens.js
@@ -1,0 +1,140 @@
+const gameScreens = [{
+  type: `tinder-like`,
+  question: `Угадай, фото или рисунок?`,
+  answers: [{
+    image: {
+      url: `https://k42.kn3.net/CF42609C8.jpg`
+    },
+    type: `painting`
+  }]
+}, {
+  type: `two-of-two`,
+  question: `Угадайте для каждого изображения фото или рисунок?`,
+  answers: [{
+    image: {
+      url: `https://k42.kn3.net/CF42609C8.jpg`
+    },
+    type: `photo`
+  },
+  {
+    image: {
+      url: `https://k42.kn3.net/CF42609C8.jpg`
+    },
+    type: `photo`
+  }
+  ]
+}, {
+  type: `tinder-like`,
+  question: `Угадай, фото или рисунок?`,
+  answers: [{
+    image: {
+      url: `https://k42.kn3.net/CF42609C8.jpg`
+    },
+    type: `photo`
+  }]
+}, {
+  type: `one-of-three`,
+  question: `Найдите рисунок среди изображений`,
+  answers: [{
+    image: {
+      url: `https://k42.kn3.net/CF42609C8.jpg`
+    },
+    type: `painting`
+  },
+  {
+    image: {
+      url: `https://k42.kn3.net/CF42609C8.jpg`
+    },
+    type: `painting`
+  },
+  {
+    image: {
+      url: `https://k42.kn3.net/CF42609C8.jpg`
+    },
+    type: `painting`
+  }
+  ]
+}, {
+  type: `tinder-like`,
+  question: `Угадай, фото или рисунок?`,
+  answers: [{
+    image: {
+      url: `https://k42.kn3.net/CF42609C8.jpg`
+    },
+    type: `painting`
+  }]
+}, {
+  type: `one-of-three`,
+  question: `Найдите рисунок среди изображений`,
+  answers: [{
+    image: {
+      url: `https://k42.kn3.net/CF42609C8.jpg`
+    },
+    type: `painting`
+  },
+  {
+    image: {
+      url: `https://k42.kn3.net/CF42609C8.jpg`
+    },
+    type: `painting`
+  },
+  {
+    image: {
+      url: `https://k42.kn3.net/CF42609C8.jpg`
+    },
+    type: `painting`
+  }
+  ]
+}, {
+  type: `two-of-two`,
+  question: `Угадайте для каждого изображения фото или рисунок?`,
+  answers: [{
+    image: {
+      url: `https://k42.kn3.net/CF42609C8.jpg`
+    },
+    type: `photo`
+  },
+  {
+    image: {
+      url: `https://k42.kn3.net/CF42609C8.jpg`
+    },
+    type: `photo`
+  }
+  ]
+}, {
+  type: `tinder-like`,
+  question: `Угадай, фото или рисунок?`,
+  answers: [{
+    image: {
+      url: `https://k42.kn3.net/CF42609C8.jpg`
+    },
+    type: `painting`
+  }]
+}, {
+  type: `two-of-two`,
+  question: `Угадайте для каждого изображения фото или рисунок?`,
+  answers: [{
+    image: {
+      url: `https://k42.kn3.net/CF42609C8.jpg`
+    },
+    type: `photo`
+  },
+  {
+    image: {
+      url: `https://k42.kn3.net/CF42609C8.jpg`
+    },
+    type: `photo`
+  }
+  ]
+}, {
+  type: `tinder-like`,
+  question: `Угадай, фото или рисунок?`,
+  answers: [{
+    image: {
+      url: `https://k42.kn3.net/CF42609C8.jpg`
+    },
+    type: `painting`
+  }]
+}];
+
+export default gameScreens;
