+++ js/render-screen.js
@@ -8,13 +8,13 @@ import stats from './screens/stats';
 import checkLives from './data/check-lives';
 
 const renderScreen = () => {
-  let lives = checkLives(answers);
-  if (!lives) {
-    return showScreen(stats(gamePlay));
-  }
-  return (gamePlay.GAME_SETUP.level < screens.length)
-    ? showScreen(screens[gamePlay.GAME_SETUP.level])
-    : showScreen(stats(gamePlay));
+  // let lives = checkLives(answers);
+  // if (!lives) {
+  //   return showScreen(stats(gamePlay));
+  // }
+  // return (gamePlay.GAME_SETUP.level < screens.length)
+  //   ? showScreen(screens[gamePlay.GAME_SETUP.level])
+  //   : showScreen(stats(gamePlay));
 };
 
 export default renderScreen;
